backend/pyproject.toml
[project]
name = "aicx-callbot"
version = "0.1.0"
description = "B2B Callbot SaaS Platform — MVP (vox 내재화)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "websockets>=13.0",
    "numpy>=1.26",
]

[project.optional-dependencies]
gcp = [
    "google-cloud-speech>=2.27",
    "google-cloud-texttospeech>=2.21",
    "google-genai>=2.0",
]
vad = [
    "silero-vad>=5.1",
    "torch>=2.2",
]
dev = [
    "ruff>=0.7",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
